telegram-downloader-plugin/requirements.txt
aiohttp
python-dotenv
telethon
tqdm
colorama